moving download of pascal into statit method
